requirements.txt
flet
pyperclip
vertexai
markdown
google-cloud-speech
pyaudio